Aqui terminou

+++ src/main/java/com/br/apidesafioverzel/adapters/in/controllers/CarController.java
@@ -3,12 +3,14 @@ package com.br.apidesafioverzel.adapters.in.controllers;
 import com.br.apidesafioverzel.application.dto.CarDTO;
 import com.br.apidesafioverzel.application.dto.CarMinDTO;
 import com.br.apidesafioverzel.application.services.CarService;
+import com.br.apidesafioverzel.application.services.UserService;
 import com.br.apidesafioverzel.domain.entities.Car;
 import jakarta.validation.Valid;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.data.domain.Page;
 import org.springframework.data.domain.Pageable;
 import org.springframework.http.ResponseEntity;
+import org.springframework.security.oauth2.server.resource.authentication.JwtAuthenticationToken;
 import org.springframework.web.bind.annotation.*;
 import org.springframework.web.servlet.support.ServletUriComponentsBuilder;
 
@@ -19,11 +21,13 @@ import java.net.URI;
 @RequestMapping("/carros")
 public class CarController {
 
-    @Autowired
-    private CarService carService;
 
-    public CarController(CarService carService) {
+    private CarService carService;
+    private UserService userService;
+
+    public CarController(CarService carService, UserService userService) {
         this.carService = carService;
+        this.userService = userService;
     }
 
     @GetMapping
@@ -40,17 +44,33 @@ public class CarController {
 
 
     @PostMapping
-    public ResponseEntity<CarDTO> insert(@Valid @RequestBody CarDTO carDTO) {
-        carDTO = carService.insert(carDTO);
-        URI uri = ServletUriComponentsBuilder.fromCurrentRequest().path("/{id}")
-                .buildAndExpand(carDTO.getId()).toUri();
-        return ResponseEntity.created(uri).body(carDTO);
+    public ResponseEntity<CarDTO> insert(@Valid @RequestBody CarDTO carDTO, JwtAuthenticationToken token) {
+
+        var user = userService.findById(Long.parseLong(token.getToken().getSubject()));
+        var isAdmin = user.getRoles().stream().anyMatch(role -> role.getName().equals("ADMIN"));
+
+        if (!isAdmin) {
+            return ResponseEntity.status(403).build();
+        } else {
+            carDTO = carService.insert(carDTO);
+            URI uri = ServletUriComponentsBuilder.fromCurrentRequest().path("/{id}")
+                    .buildAndExpand(carDTO.getId()).toUri();
+            return ResponseEntity.created(uri).body(carDTO);
+        }
+
     }
 
     @DeleteMapping("/{id}")
-    public ResponseEntity<Void> delete(@PathVariable Long id) {
-        carService.delete(id);
-        return ResponseEntity.noContent().build();
+    public ResponseEntity<Void> delete(@PathVariable Long id, JwtAuthenticationToken token) {
+        var user = userService.findById(Long.parseLong(token.getToken().getSubject()));
+        var isAdmin = user.getRoles().stream().anyMatch(role -> role.getName().equals("ADMIN"));
+
+        if (!isAdmin) {
+            return ResponseEntity.status(403).build();
+        } else {
+            carService.delete(id);
+            return ResponseEntity.noContent().build();
+        }
     }
